efw-runtime: code refactoring for runtime module
This commit obsoletes internal runtime module by moving runtime
structure to the top of runtime crate.
