+++ services/src/bin/snd-fireworks-ctl-service.rs
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // Copyright (c) 2020 Takashi Sakamoto
 use snd_firewire_ctl_services::*;
-use efw_runtime::runtime::EfwRuntime;
+use efw_runtime::EfwRuntime;
 
 struct EfwServiceCmd;
